refactor: execute task
