+++ package.json
@@ -50,6 +50,11 @@
     "verify:bytecodes": "yarn ts-node ./bin/verify-bytecodes.ts",
     "deploy": "./bin/deploy.sh",
     "deploy:localhost": "yarn hardhat deploy --network localhost",
+    "deploy:fork" :"yarn hardhat deploy",
+    "deploy:alfajores":"yarn hardhat deploy --network alfajores",
+    "deploy:baklava":"yarn hardhat deploy --network baklava",
+    "deploy:celo":"yarn hardhat deploy --network celo",
+    "execute": "yarn hardhat executeProposal",
     "build": "./bin/build.sh",
     "script:dev": "./bin/dev-script.sh",
     "cgp": "./bin/cgp.sh",
